Refact plate
扱う型を最小限にしたけど意味なかった

+++ include/VirtualPlate.hpp
@@ -10,8 +10,8 @@ enum class Stone {
   Blue
 };
 
-static const int dx[] = { 1, 1, 1, 0 };
-static const int dy[] = {-1, 0, 1, 1 };
+static const short dx[] = { 1, 1, 1, 0 };
+static const short dy[] = {-1, 0, 1, 1 };
 
 class VirtualPlate {
  public:
@@ -21,14 +21,14 @@ class VirtualPlate {
   ~VirtualPlate() = default;
   void init();
   void copy_plate(const VirtualPlate& src);
-  void insert(int input_x);
-  inline bool can_drop(int x, int y) const;
+  void insert(short input_x);
+  inline bool can_drop(short x, short y) const;
   void switch_active_stone();
   bool can_continue() const;
-  inline bool is_inside_plate(int x, int y) const;
+  inline bool is_inside_plate(short x, short y) const;
   bool is_game_finish() const;
-  int get_length(int x, int y, int dx, int dy) const;
-  bool is_valid_hand(int x) const;
+  int get_length(short x, short y, short dx, short dy) const;
+  bool is_valid_hand(short x) const;
   friend void show(VirtualPlate game_plate);
   Stone get_active_stone() const;
 private:
